LVGL on ESP32, and all three slices merged and verified together
The pet has a screen. ADR 0027 brings LVGL up as an ESP-IDF component
compiling the portable glue out of simulator/src/lvgl/ by relative path,
and app_main.cpp now drives kf_lvgl_port_init()/kf_screen_nav_init() at
boot and pet session -> screen nav -> LVGL pump each frame, the same
order sdl_main.cpp uses so a screen never shows last frame's numbers.
KF_DEMO_SPRITE becomes KF_DEMO_NONE: LVGL owns the framebuffer now, and
the demo sprite would fight it for the same pixels.

This commit is mostly the merge. Two parallel tasks and this branch all
edited ports/esp32/CMakeLists.txt, main/CMakeLists.txt and app_main.cpp;
the conflicts were resolved by hand rather than by taking a side, and
one factual error was corrected in the process. The LVGL task's version
of the header comment claimed kf_time_wall() has no DS3231 behind it and
that this port had not been switched to one. That was true when ADR 0025
was written and is not true now: ADR 0026 landed the driver, and the
bring-up diagnostic has since confirmed the clock advancing across a
genuine power cut on coin-cell power. The merged comment says so.

ADR numbering reconciled while it was cheap. The Lua task picked 0030
defensively, assuming concurrent slices had taken 0027-0029; only 0027
was actually claimed. Lua is renumbered 0028 and the new panel-profile
ADR takes 0029, so the sequence has no holes and the index is honest.
This is the same class of problem as the 0023 duplicate fixed earlier
today, and it is much cheaper to fix before the numbers are referenced
than after.

Adds ADR 0029, documenting the panel-profile decision from 22921e6 --
including the alternatives rejected, particularly making kf_color
big-endian at construction, which is faster and survives a first look
but silently invalidates every pre-baked sprite in flash.

Verified together, not just individually: full idf.py fullclean +
build clean for esp32s3 against ESP-IDF v6.0.2, zero warnings, firmware
0x98f80 bytes with 40% of the app partition free. Desktop headless
builds clean on macOS/clang, all 13 ctest cases pass, and
check_no_heap.py still reports the core heap-free.

NOT verified on hardware. Nothing in this commit has been flashed. The
pet screen has never been seen on glass.

+++ simulator/src/lvgl/kf_lvgl_display.cpp
@@ -69,7 +69,16 @@ lv_display_t *kf_lvgl_display_init(void) {
      * nothing to do with it. */
     lv_display_set_theme(disp, lv_theme_simple_init(disp));
 
+    /* static_cast<int>(kDrawBufRows), not kDrawBufRows bare: found on the
+     * ESP32 target, not desktop -- xtensa-esp32s3-elf's <cstdint> typedefs
+     * int32_t to `long`, a distinct type from `int` even though both are
+     * 32 bits there, so %d (which only ever promises `int`) genuinely
+     * mismatches kDrawBufRows's declared type on that one target. Desktop's
+     * int32_t is plain `int`, so this was never visible there. Casting
+     * makes the actual printf argument type match %d's contract on every
+     * target, not just the one that happened to warn first. */
     KF_LOGI(TAG, "%dx%d RGB565, %d-row partial draw buffer (%zu bytes)",
-            KF_DISPLAY_WIDTH, KF_DISPLAY_HEIGHT, kDrawBufRows, kDrawBufBytes);
+            KF_DISPLAY_WIDTH, KF_DISPLAY_HEIGHT,
+            static_cast<int>(kDrawBufRows), kDrawBufBytes);
     return disp;
 }
